Moved the sentry tile component to the `Http\Livewire` folder. Made some improvement so we keep the dynamic functionality of livewire.
